+++ src/SharpEmu.HLE/Host/Windows/WindowsHostPlatform.cs
@@ -10,4 +10,6 @@ internal sealed class WindowsHostPlatform : IHostPlatform
     public IHostThreading Threading { get; } = new WindowsHostThreading();
 
     public IHostSymbolResolver Symbols { get; } = new WindowsHostSymbolResolver();
+
+    public IHostAudioOutput Audio { get; } = new WindowsWaveOutAudio();
 }
